Split Client into separate module
